Moved file location to widgets
